Patch to create MS linker compatible import libraries.
These patches, when applied to the binutils tree, will allow the
linker to create import libraries that are compatible with the
MS linker.

Note that dlltool will not generate these types of libraries; this
patch only applies to the linker.

longname-fix.patch fixes the long name table to be NULL terminated,
which is a deviation of the PE-COFF spec.

implib-compat.patch fixes the generation of the archive member
names so that they are the same, as the MS linker will not properly
combine the import sections otherwise.